CORE-302 Remove unused ABI
